split up client files
