+++ modules/client/src/router.tsx
@@ -2,11 +2,9 @@ import React, { ReactElement } from 'react';
 import { BrowserRouter, Switch, Route } from 'react-router-dom';
 import { Navigation } from './body';
 import Home from './routes/home/index';
-import {
-  PlayerList,
-  PlayerDetails,
-  PlayerFormLoader,
-} from './routes/players/index';
+import PlayerForm from './routes/players/form';
+import PlayerList from './routes/players/list';
+import PlayerDetails from './routes/players/details';
 import { CourseList, CourseDetails } from './routes/courses/index';
 import { EventList, EventDetails } from './routes/events/index';
 import { SeasonList, SeasonDetails } from './routes/seasons/index';
@@ -18,8 +16,8 @@ const Router = (): ReactElement => {
         <Switch>
           <Route exact path="/" component={Home} />
           <Route exact path="/players" component={PlayerList} />
-          <Route exact path="/players/new" component={PlayerFormLoader} />
-          <Route exact path="/players/edit/:id" component={PlayerFormLoader} />
+          <Route exact path="/players/new" component={PlayerForm} />
+          <Route exact path="/players/edit/:id" component={PlayerForm} />
           <Route exact path="/players/:id" component={PlayerDetails} />
           <Route exact path="/courses" component={CourseList} />
           <Route exact path="/courses/:id" component={CourseDetails} />
