+++ modules/client/src/routes/types.ts
@@ -1,0 +1,3 @@
+export interface EntityDetailsParams {
+  id: string;
+}
